# end2endPoC/kogito-consumer/app/target/generatedtest/openapi.json
{"openapi":"3.0.1","paths":{"/executions":{"get":{"operationId":"getExecutions","requestBody":{"content":{"*/*":{"schema":{"type":"string"}}}},"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecutionResponse"}}}}}}},"/executions/{key}":{"get":{"operationId":"getExecutionByKey","requestBody":{"content":{"*/*":{"schema":{"type":"string"}}}},"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DecisionDetailResponse"}}}}}}},"/executions/{key}/inputs":{"get":{"operationId":"getExecutionInputs","requestBody":{"content":{"*/*":{"schema":{"type":"string"}}}},"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DecisionInputsResponse"}}}}}}},"/executions/{key}/outcomes":{"get":{"operationId":"getExecutionOutcome","requestBody":{"content":{"*/*":{"schema":{"type":"string"}}}},"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DecisionOutputsResponse"}}}}}}},"/models":{"get":{"operationId":"getModels","responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailableModelsResponse"}}}}}}},"/models/{id}":{"get":{"operationId":"getModelById","requestBody":{"content":{"*/*":{"schema":{"type":"string"}}}},"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelResponse"}}}}}}},"/models/{id}/evaluate":{"post":{"operationId":"evaluateDecision","requestBody":{"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluationResponse"}}}}}}},"/{job-id}":{"get":{"operationId":"get","parameters":[{"name":"wait","in":"query","schema":{"type":"integer","format":"int64","default":-1}},{"name":"job-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"default":{"description":"default response","content":{"*/*":{}}}}},"post":{"operationId":"readAndRemove","parameters":[{"name":"wait","in":"query","schema":{"type":"integer","format":"int64","default":-1}},{"name":"job-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"default":{"description":"default response","content":{"*/*":{}}}}},"delete":{"operationId":"remove","parameters":[{"name":"job-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"default":{"description":"default response","content":{"*/*":{}}}}}}},"components":{"schemas":{"ExecutionHeaderResponse":{"type":"object","properties":{"executionId":{"type":"string"},"executionDate":{"type":"string","format":"date-time"},"executionSucceeded":{"type":"boolean"},"executorName":{"type":"string"},"executionType":{"type":"string","enum":["DECISION","PROCESS"]}}},"ExecutionResponse":{"type":"object","properties":{"total":{"type":"integer","format":"int32"},"limit":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int32"},"headers":{"type":"array","items":{"$ref":"#/components/schemas/ExecutionHeaderResponse"}}}},"DecisionDetailResponse":{"type":"object","properties":{"header":{"$ref":"#/components/schemas/ExecutionHeaderResponse"}}},"DecisionInputsResponse":{"type":"object","properties":{"input":{"type":"object"},"header":{"$ref":"#/components/schemas/ExecutionHeaderResponse"}}},"DecisionOutputsResponse":{"type":"object","properties":{"decisions":{"type":"array","items":{"$ref":"#/components/schemas/DecisionResultModel"}},"header":{"$ref":"#/components/schemas/ExecutionHeaderResponse"}}},"DecisionResultModel":{"type":"object","properties":{"decisionId":{"type":"string"},"decisionName":{"type":"string"},"evaluationStatus":{"type":"string","enum":["NOT_EVALUATED","EVALUATING","SUCCEEDED","SKIPPED","FAILED"]},"decisionResult":{"type":"string"},"messages":{"type":"array","items":{"type":"string"}},"hasErrors":{"type":"boolean"}}},"AvailableModelsResponse":{"type":"object","properties":{"availableModels":{"type":"array","items":{"$ref":"#/components/schemas/ModelDetail"}}}},"ModelDetail":{"type":"object","properties":{"modelNamespace":{"type":"string"},"modelName":{"type":"string"},"modelVersion":{"type":"string"},"creationDate":{"type":"string"}}},"ModelResponse":{"type":"object","properties":{"metainformation":{"$ref":"#/components/schemas/ModelDetail"},"inputStructure":{"type":"object"},"decisions":{"type":"array","items":{"type":"object"}}}},"EvaluationResponse":{"type":"object","properties":{"decisions":{"type":"array","items":{"type":"object"}}}},"EvaluationRequestBody":{"type":"object","properties":{"input":{"type":"object"}}}}}}